match android

+++ Source/iOS/Harvest/Harvest/Model/Information/InfoRepresentation.swift
@@ -128,7 +128,7 @@ extension Worker {
     let orchards = Entities.shared.orchards
     
     let orchardSection = SelectableSection<ListCheckRow<Orchard>>(
-      "Assigned Orchard",
+      "Assigned Orchards",
       selectionType: .multipleSelection)
     
     for (_, orchard) in orchards {
@@ -246,8 +246,15 @@ extension Worker {
     let deleteWorkerRow = ButtonRow { row in
       row.title = "Delete Worker"
     }.onCellSelection { (_, _) in
-      HarvestDB.delete(worker: self) { (_, _) in
-        formVC.navigationController?.popViewController(animated: true)
+      UIAlertController.present(
+        question: "Are You Sure You Want to Delete \(self.name)?",
+        detail: """
+        You will not be able to get back any information about this worker.
+        Any work done by this worker will no longer have any statistics associated with them.
+        """, on: formVC) {
+          HarvestDB.delete(worker: self) { (_, _) in
+            formVC.navigationController?.popViewController(animated: true)
+          }
       }
     }.cellUpdate { (cell, _) in
       cell.textLabel?.textColor = .white
@@ -271,7 +278,9 @@ extension Worker {
       <<< lastnameRow
       <<< idRow
       
-      +++ Section("Role")
+      +++ Section.init(
+        header: "Role",
+        footer: "Any foreman phone number must include its area code. Example +27 71 1234567")
       <<< isForemanRow
       <<< phoneRow
 //      <<< emailRow
@@ -284,7 +293,7 @@ extension Worker {
     _ = id != "" ? (form +++ performanceSection) : form
     
     form
-      +++ Section("Information")
+      +++ Section("Further Information")
       <<< infoRow
       
       +++ Section()
@@ -403,8 +412,14 @@ extension Farm {
     let deleteFarmRow = ButtonRow { row in
       row.title = "Delete Farm"
     }.onCellSelection { (_, _) in
-      HarvestDB.delete(farm: self) { (_, _) in
-        formVC.navigationController?.popViewController(animated: true)
+      UIAlertController.present(
+        question: "Are You Sure You Want to Delete \(self.name)?",
+        detail: """
+        You will not be able to get back any information about this farm.
+        """, on: formVC) {
+          HarvestDB.delete(farm: self) { (_, _) in
+            formVC.navigationController?.popViewController(animated: true)
+          }
       }
     }.cellUpdate { (cell, _) in
       cell.textLabel?.textColor = .white
@@ -426,7 +441,7 @@ extension Farm {
       +++ orchardsSection
       <<< orchardRow
       
-      +++ Section("Information")
+      +++ Section("Further Information")
       <<< detailsRow
     
       +++ Section()
@@ -599,14 +614,15 @@ extension Orchard {
       onChange()
     }
     
-    let farmSelection = PickerRow<Farm> { row in
+    let farmSelection = PushRow<Farm> { row in
+      row.title = "Assigned Farm"
       row.add(rule: RuleRequired(msg: "• A farm must be selected for an orchard to be a part of"))
       row.options = []
       var aFarm: Farm? = nil
       
       for (_, farm) in farms {
         let farm = farm
-        row.options.append(farm)
+        row.options?.append(farm)
         if farm.id == assignedFarm {
           aFarm = farm
         }
@@ -640,12 +656,51 @@ extension Orchard {
     let deleteOrchardRow = ButtonRow { row in
       row.title = "Delete Orchard"
     }.onCellSelection { (_, _) in
-      HarvestDB.delete(orchard: self) { (_, _) in
-        formVC.navigationController?.popViewController(animated: true)
+      UIAlertController.present(
+        question: "Are You Sure You Want to Delete \(self.name)?",
+        detail: """
+        You will not be able to get back any information about this farm.
+        Any work done in this orchard will no longer display any statistics.
+        """, on: formVC) {
+          HarvestDB.delete(orchard: self) { (_, _) in
+            formVC.navigationController?.popViewController(animated: true)
+          }
       }
     }.cellUpdate { (cell, _) in
       cell.textLabel?.textColor = .white
       cell.backgroundColor = .red
+    }
+    
+    let assignedWorkersRow = MultipleSelectorRow<Worker> { row in
+      row.title = "Assigned Workers"
+      self.tempory?.assignedWorkers = []
+      var opts = [Worker]()
+      var vals = Set<Worker>()
+      for (_, worker) in Entities.shared.workers {
+        opts.append(worker)
+        if worker.assignedOrchards.contains(id) {
+          vals.insert(worker)
+          self.tempory?.assignedWorkers.append((worker.id, .assigned))
+        } else {
+          self.tempory?.assignedWorkers.append((worker.id, .unassigned))
+        }
+        row.options = opts
+        row.value = vals
+      }
+    }.onChange { row in
+      var idx = 0
+      for (assignedWorker, status) in self.tempory?.assignedWorkers ?? [] {
+        if [.remove, .unassigned].contains(status)
+        && row.value?.contains(where: { $0.id == assignedWorker }) ?? false {
+          self.tempory?.assignedWorkers[idx] = (assignedWorker, .add)
+        } else if [.add, .assigned].contains(status)
+        && !(row.value?.contains(where: { $0.id == assignedWorker }) ?? false) {
+          self.tempory?.assignedWorkers[idx] = (assignedWorker, .remove)
+        }
+        
+        idx += 1
+      }
+      onChange()
     }
     
     form
@@ -669,15 +724,18 @@ extension Orchard {
       <<< widthRow
       <<< heightRow
     
-      +++ Section("Part of Farm")
+      +++ Section("Assigned Farm")
       <<< farmSelection
+    
+      +++ Section("Assigned Workers")
+      <<< assignedWorkersRow
       
     if id != "" {
       form +++ performanceSection
     }
     
     form
-      +++ Section("Information")
+      +++ Section("Further Information")
       <<< detailsRow
       
       +++ Section()
@@ -742,14 +800,14 @@ extension HarvestUser {
     let form = formVC.form
     temporary = HarvestUser(json: json())
     
-    let farmNameRow = TextRow { row in
-      row.title = "Farm Name"
-      row.value = HarvestUser.current.farmName
-      row.placeholder = "Name of the farm"
+    let organisationNameRow = TextRow { row in
+      row.title = "Organisation Name"
+      row.value = HarvestUser.current.organisationName
+      row.placeholder = "Name of the Organisation"
     }.cellUpdate { (cell, _) in
       cell.textField.clearButtonMode = .whileEditing
     }.onChange { row in
-      self.temporary?.farmName = row.value ?? ""
+      self.temporary?.organisationName = row.value ?? ""
       onChange()
     }
     
@@ -776,8 +834,8 @@ extension HarvestUser {
     }
     
     form
-      +++ Section("Farm Name")
-      <<< farmNameRow
+      +++ Section("Organisation Name")
+      <<< organisationNameRow
       
       +++ Section("Farmer Name")
       <<< firstnameRow
